Move remaining DAO stuff to es instance and delete dao class.
